refactor : 비즈니스 로직을 hook, utils 등으로 분리
