+++ src/features/chat/model/constants.ts
@@ -1,0 +1,12 @@
+export const GRADIENTS = [
+  'bg-gradient-to-r from-purple-400 via-pink-500 to-orange-500',
+  'bg-gradient-to-r from-blue-400 via-cyan-400 to-teal-400',
+  'bg-gradient-to-r from-red-400 via-yellow-400 to-green-400',
+  'bg-gradient-to-r from-fuchsia-400 via-pink-400 to-rose-400',
+  'bg-gradient-to-r from-indigo-400 via-sky-400 to-emerald-400',
+  'bg-gradient-to-r from-orange-400 via-yellow-400 to-green-400',
+  'bg-gradient-to-r from-purple-400 via-indigo-400 to-pink-500',
+  'bg-gradient-to-r from-red-400 via-orange-400 to-yellow-400',
+  'bg-gradient-to-r from-green-400 via-teal-400 to-cyan-400',
+  'bg-gradient-to-r from-blue-400 via-sky-400 to-emerald-400',
+];
